18-11-2021

+++ app/src/main/java/com/example/e_commerce/models/auth/login/Data.kt
@@ -1,12 +1,12 @@
 package com.example.e_commerce.models.auth.login
 
 data class Data(
-    val credit: Int,
+    val credit: Double,
     val email: String,
     val id: Int,
     val image: String,
     val name: String,
     val phone: String,
-    val points: Int,
+    val points: Double,
     val token: String
 )
